datacentric: add some detail metrics
Adding some additional information from intel load latency facility, and
rearrange metric column position so that we display the most significant
metrics.

+++ src/lib/analysis/CallPath.cpp
@@ -166,7 +166,7 @@ read(const Util::StringVec& profileFiles, const Util::UIntVec* groupMap,
     // add the directory into the set of directories
     prof->addDirectory(profileFiles[i]);
   }
-  prof->metricMgr()->mergePerfEventStatistics_finalize(profileFiles.size());
+  prof->metricMgr()->mergePerfEventStatistics_finalize((int)profileFiles.size());
   
   return prof;
 }
